Update rmatmul for AVX2 and AVX512
